Move the NotChecker to its own file
